backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx
python-multipart==0.0.6
RestrictedPython==8.0
pandas>=2.2.0
langtrace-python-sdk
pytest>=8.2.0
pytest-asyncio>=0.26.0
aiofiles==23.2.1
rouge-score>=0.1.2
nltk>=3.9.1
datasets>=3.5.1
evaluate>=0.4.3
python-dotenv==1.0.0
requests>=2.32.2
openpyxl==3.1.2
openai>=1.7.1,<2.0.0
pydantic>=2.4.2,<3.0.0
psutil>=5.9.8
asyncpg>=0.29.0
supabase>=2.15.1
crewai==0.5.0
langchain>=0.0.0
langchain-core>=0.0.0
rich>=13.7.0
traceback-with-variables>=2.0.4
PyPDF2>=3.0.0
python-docx>=0.8.11
sentence-transformers>=2.2.2
anthropic>=0.18.0
